+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/model/TyeServiceMetric.kt
@@ -1,0 +1,3 @@
+package com.github.rafaelldi.tyeplugin.model
+
+data class TyeServiceMetric(val name: String, val instance: String?, val value: String?)