+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/remoteServer/TyeHostConfiguration.kt
@@ -5,6 +5,7 @@ import io.ktor.http.*
 
 class TyeHostConfiguration : CloudConfigurationBase<TyeHostConfiguration>() {
     var hostAddress = "http://localhost:8000"
+    var monitorMetrics = false
 
     companion object {
         fun createDefault(): TyeHostConfiguration {